@@ -14,20 +14,20 @@
         "php": "^8.0",
         "algolia/algoliasearch-client-php": "3.2.0",
         "appstract/laravel-opcache": "4.0.2",
-        "aws/aws-sdk-php": "3.215.2",
+        "aws/aws-sdk-php": "3.218.2",
         "biscolab/laravel-recaptcha": "5.3.0",
         "bugsnag/bugsnag-laravel": "2.23.0",
         "curl/curl": "2.3.3",
-        "doctrine/dbal": "3.3.3",
+        "doctrine/dbal": "3.3.4",
         "geoip2/geoip2": "2.12.2",
         "google/recaptcha": "1.2.4",
         "guzzlehttp/guzzle": "7.4.2",
         "intervention/image": "2.7.1",
         "josiasmontag/laravel-localization": "1.0.0",
-        "laravel/framework": "9.5.1",
+        "laravel/framework": "9.6.0",
         "laravel/slack-notification-channel": "2.4.0",
         "laravel/socialite": "5.5.2",
-        "laravel/tinker": "2.7.1",
+        "laravel/tinker": "2.7.2",
         "maennchen/zipstream-php": "2.1.0",
         "mattkingshott/quest": "2.0.1",
         "sentry/sentry-laravel": "2.11.1",
@@ -35,9 +35,9 @@
         "socialiteproviders/google": "4.1.0",
         "spatie/array-to-xml": "3.1.2",
         "spatie/eloquent-sortable": "4.0.1",
-        "spatie/laravel-fractal": "6.0.0",
+        "spatie/laravel-fractal": "6.0.2",
         "sunra/php-simple-html-dom-parser": "1.5.2",
-        "symfony/http-client": "6.0.5",
+        "symfony/http-client": "6.0.7",
         "symfony/psr-http-message-bridge": "2.1.2",
         "symfony/yaml": "6.0.3",
         "torann/geoip": "3.0.4",
@@ -53,7 +53,7 @@
         "itsgoingd/clockwork": "5.1.5",
         "laravel/legacy-factories": "1.3.0",
         "mockery/mockery": "1.5.0",
-        "phpunit/phpunit": "9.5.19",
+        "phpunit/phpunit": "9.5.20",
         "spatie/laravel-tail": "4.4.0",
         "squizlabs/php_codesniffer": "3.6.2"
     },
